requirements.txt
streamlit>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0 